+++ frontend/wailsjs/go/main/App.js
@@ -78,6 +78,14 @@ export function ScanAssets() {
   return window['go']['main']['App']['ScanAssets']();
 }
 
+export function SelectDirectory(arg1) {
+  return window['go']['main']['App']['SelectDirectory'](arg1);
+}
+
+export function SelectFile(arg1) {
+  return window['go']['main']['App']['SelectFile'](arg1);
+}
+
 export function SelectHashcatBinary() {
   return window['go']['main']['App']['SelectHashcatBinary']();
 }